+++ src/taoensso/carmine/tundra/s3.clj
@@ -32,7 +32,7 @@
          (s3/create-bucket creds bucket)
          (recur k v))
        (instance? Exception reply) reply
-       :else (Exception. (format "Unexpected reply: %s" reply)))))
+       :else (ex-info (format "Unexpected reply: %s" reply) {:reply reply}))))
 
   (fetch-keys [this ks]
     (let [fetch1